clean up code

+++ src/main/scala/net/houzuo/aurinko2/io/SimpleIO.scala
@@ -4,6 +4,7 @@ import java.io.BufferedWriter
 import java.io.File
 import java.io.FileWriter
 import java.io.IOException
+
 import scala.collection.mutable.ListBuffer
 
 object SimpleIO {
@@ -11,9 +12,8 @@ object SimpleIO {
   /** Overwrite or append the lines to the destination file. Create the file if it does not exist. */
   def spit(filename: String, lines: Seq[String], append: Boolean = false) {
     val file = new File(filename)
-    if (!file.exists())
-      if (!file.createNewFile())
-        throw new IOException(s"Cannot create file $filename")
+    if (!file.exists() && !file.createNewFile())
+      throw new IOException(s"Cannot create file $filename")
 
     if (!file.canWrite())
       throw new IOException(s"You do not have permission to write to $filename")
@@ -27,10 +27,9 @@ object SimpleIO {
   def rmrf(dir: File) {
     val failures = new ListBuffer[String]
 
-    if (dir.isFile()) {
-      if (!dir.delete())
-        failures += dir.getAbsolutePath()
-    } else {
+    if (dir.isFile() && !dir.delete())
+      failures += dir.getAbsolutePath()
+    else {
       for (file <- dir.listFiles())
         rmrf(file)
       dir.delete()
